Delete user now returns boolean if transaction was successful, update user now copies changed properties straight into the db user, add user implements interface method

+++ src/main/java/java_advance/spring_boot_spotify/service/UserService.java
@@ -2,16 +2,15 @@ package java_advance.spring_boot_spotify.service;
 
 import java_advance.spring_boot_spotify.model.User;
 import java_advance.spring_boot_spotify.repository.UserRepository;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 @Service
 public class UserService implements UserServiceInterface {
-    private static final int FIRST_NAME_INDEX = 0;
-    private static final int LAST_NAME_INDEX = 1;
-    private static final int EMAIL_INDEX = 2;
-    private static final int PHONE_INDEX = 3;
+
+    private static final String[] IGNORED_PROPERTIES_FOR_USER_UPDATE = {"playlists", "archived"};
 
     private UserRepository userRepository;
 
@@ -28,31 +27,32 @@ public class UserService implements UserServiceInterface {
     @Override
     public List<User> getAllUsers() {
         return userRepository.findAll();
+
     }
 
     @Override
-    public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
+    public boolean deleteUserById(Long id) {
+        if (!(getUserById(id) == null)) {
+            userRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     @Override
-    public User updateUserById(Long id, List<String> userDetails) {
-        User user = userRepository.findById(id).orElse(null);
-        user.setFirstName(userDetails.get(FIRST_NAME_INDEX));
-        user.setLastName(userDetails.get(LAST_NAME_INDEX));
-        user.setEmail(userDetails.get(EMAIL_INDEX));
-        user.setPhone(userDetails.get(PHONE_INDEX));
-
-        return userRepository.save(user);
+    public User updateUserById(User providedUpdatedUser) {
+        User dbUser = userRepository.findById(providedUpdatedUser.getUserId()).orElse(null);
+        if (dbUser == null) {
+            return null;
+        } else {
+            BeanUtils.copyProperties(providedUpdatedUser, dbUser, IGNORED_PROPERTIES_FOR_USER_UPDATE);
+        }
+        return userRepository.save(dbUser);
     }
 
     @Override
-    public User addUser(List<String> userDetails) {
-        User user = new User(userDetails.get(FIRST_NAME_INDEX),
-                            userDetails.get(LAST_NAME_INDEX),
-                            userDetails.get(EMAIL_INDEX),
-                            userDetails.get(PHONE_INDEX));
-
-        return userRepository.save(user);
+    public User addUser(User newUser) {
+        return userRepository.save(newUser);
     }
 }
